Renamed card to card.ng1
